Architectural improvements and code optimizations
