+++ src/cedict-service/store.js
@@ -1,179 +1,156 @@
+/**
+ * @module Store
+ */
+
+/** A base class for data stores */
 export default class Store {
-  constructor (schema) {
-    Store.checkSchemaValidity(schema)
-    this._schema = schema
-    // DB object is not available when store object is created. It must be added later.
-    this._db = null
+  constructor (configuration) {
+    this.constructor.checkConfiguration(configuration)
+    this._configuration = configuration
   }
 
   /**
-   * Checks if the schema supplied has all the necessary information in it.
-   * If schema is not valid it will throw an error indicating which check failed.
+   * Checks if the configuration supplied has all the necessary information in it.
+   * If configuration is not valid it will throw an error indicating which check failed.
    *
-   * @param {object} schema - A JSON like schema object.
+   * @param {object} configuration - A JSON like configuration object.
    */
-  static checkSchemaValidity (schema) {
-    if (!schema.name) throw new Error('A store name is missing from a schema')
-    if (!schema.primaryIndex) throw new Error('A primaryIndex tree is missing from a schema')
+  static checkConfiguration (configuration) {
+    if (!configuration.name) throw new Error('A store name is missing from a configuration')
   }
 
-  associateWith (db) {
-    this._db = db
+  /**
+   * Associates a store with a container where it exists: a remote storage, an IndexedDB, etc.
+   *
+   * @param {object} storeObject - A store object.
+   * @returns {Store} A self-reference for chaining.
+   */
+  associateWith (storeObject) {
     return this
   }
 
   /**
-   * This method can be run only from `onupgradeneeded` callback
+   * Creates a store. Can be run from `onupgradeneeded` callback only.
+   * Must be implemented in a subclass.
+   *
+   * @returns {Promise<undefined>|Promise<Error>} A promise that is resolved if store was created successfully
+   *          and is rejected if creation failed.
    */
   create () {
-    return new Promise((resolve, reject) => {
-      const options = this._schema.primaryIndex.keyPath ? { keyPath: this._schema.primaryIndex.keyPath } : undefined
-      console.info(`${this._schema.name} store create`, options)
-      const store = this._db.createObjectStore(this._schema.name, options)
-      if (this._schema.indexes) {
-        Object.values(this._schema.indexes).forEach(idx => {
-          console.info('Creating an index for', idx)
-          try {
-            store.createIndex(idx.name, idx.keyPath, { unique: idx.unique })
-          } catch (error) {
-            reject(error)
-          }
-        })
-      }
+    return new Promise((resolve) => {
+      resolve()
     })
   }
 
+  /**
+   * Deletes all records from the store.
+   * Must be implemented in a subclass.
+   *
+   * @returns {Promise<undefined>|Promise<Error>} A promise that is resolved if all records were removed successfully
+   *          and is rejected if operation failed.
+   */
   clear () {
-    let transaction = this._db.transaction(this._schema.name, Store.accessModes.READ_WRITE) // eslint-disable-line prefer-const
-
-    // create an object store on the transaction
-    let objectStore = transaction.objectStore(this._schema.name) // eslint-disable-line prefer-const
-
-    // Make a request to clear all the data out of the object store
-    let objectStoreRequest = objectStore.clear() // eslint-disable-line prefer-const
-
-    objectStoreRequest.onsuccess = (event) => {
-      // report the success of our request
-      console(`${this._schema.name}: clear request success`, event)
-    }
-
-    // report on the success of the transaction completing, when everything is done
-    transaction.oncomplete = (event) => {
-      console(`${this._schema.name}: clear transaction has been completed`, event)
-    }
-
-    transaction.onerror = (event) => {
-      console(`${this._schema.name}: clear transaction error`, event)
-    }
+    return new Promise((resolve) => {
+      resolve()
+    })
   }
 
+  /**
+   * Deletes a store from its container.
+   * Must be implemented in a subclass.
+   *
+   * @returns {Promise<undefined>|Promise<Error>} A promise that is resolved if store was removed successfully
+   *          and is rejected if operation failed.
+   */
   destroy () {
-    try {
-      this._db.deleteObjectStore(this._schema.name)
-    } catch (error) {
-      console.error(error)
-    }
+    return new Promise((resolve) => {
+      resolve()
+    })
   }
 
-  // All records for a specific key
+  /**
+   * Retrieves all records from the store for a single key. Options object is implementation specific.
+   * Must be implemented in a subclass.
+   *
+   * @param {any} key - A key that specifies which records to retrieve.
+   * @param {object} options - Additional configuration parameters.
+   * @returns {Promise<object[]>|Promise<Error>} A promise that is resolved with an array of records if records
+   *          exist in a store or with an empty array if not. A promise rejection is returned if operation failed.
+   */
   get (key, options) {
-    return this.getEntries([key], options).then((result) => result[key])
+    return new Promise((resolve) => {
+      resolve()
+    })
   }
 
-  // Return all records for specific keys
+  /**
+   * Retrieves all records from the store for one or several keys. Options object is implementation specific.
+   * Must be implemented in a subclass.
+   *
+   * @param {any|any[]} keys - A key or an array of keys that specifies which records to retrieve.
+   * @param {object} options - Additional configuration parameters.
+   * @returns {Promise<{key: object[]}>|Promise<Error>} A promise that is resolved with an object. If contains keys
+   *          as its properties and values are arrays of records.
+   *          A promise rejection is returned if operation failed.
+   */
   getEntries (keys, { index = '' } = {}) {
-    return new Promise((resolve, reject) => {
-      if (!this._db) reject(new Error('Database object is missing'))
-      if (!keys) keys = reject(new Error('No keys provided'))
-      if (!Array.isArray(keys)) keys = [keys]
-      const transaction = this._db.transaction(this._schema.name, Store.accessModes.READ)
-      const store = transaction.objectStore(this._schema.name)
-      // Create an object with keys as its props
-      let result = keys.reduce((accumulator, key) => { accumulator[key] = []; return accumulator }, {}) // eslint-disable-line prefer-const
-      console.info('getEntries result is:', result)
-      /*
-      The order of request execution is guaranteed in IndexedDB.
-      This means that when the last request is completed all previous requests are done too.
-       */
-      for (let i = 0; i < keys.length; i++) {
-        const key = keys[i]
-        let getRequest
-        if (index) {
-          // Use index to retrieve a record
-          console.info(`Retrieving using a ${index} index`)
-          const dbIndex = store.index(index)
-          getRequest = dbIndex.getAll(IDBKeyRange.only(key))
-        } else {
-          console.info('Retrieving without index')
-          getRequest = store.getAll(IDBKeyRange.only(key))
-        }
-        getRequest.onsuccess = () => {
-          const records = getRequest.result
-          console.info(`(${i}) Records returned for ${key} key are:`, records)
-          result[key] = getRequest.result
-          if (i === keys.length - 1) {
-            // A last request has been completed
-            resolve(result)
-          }
-        }
-      }
-      // Transaction is completed later than `getRequest.onsuccess` is triggered
-      transaction.oncomplete = () => console.info('getAll transaction is complete')
-      transaction.onerror = (error) => { console.info('getAll transaction error'); reject(error) }
+    return new Promise((resolve) => {
+      resolve()
     })
   }
 
+  /**
+   * Retrieves all records that exist in the store.
+   * Must be implemented in a subclass.
+   *
+   * @returns {Promise<object[]>|Promise<Error>} A promise that is resolved with an array of records if records
+   *          exist in a store or with an empty array if not. A promise rejection is returned if operation failed.
+   */
   getAllEntries () {
-    return new Promise((resolve, reject) => {
-      if (!this._db) reject(new Error('Database object is missing'))
-      const transaction = this._db.transaction(this._schema.name, Store.accessModes.READ)
-      const store = transaction.objectStore(this._schema.name)
-      const getRequest = store.getAll()
-      getRequest.onsuccess = () => {
-        const records = getRequest.result
-        console.info('Records returned are:', records)
-        resolve(records)
-      }
-      // Transaction is completer later than `getRequest.onsuccess` is triggered
-      transaction.oncomplete = () => console.info('getAll transaction is complete')
-      transaction.onerror = (error) => { console.info('getAll transaction error'); reject(error) }
+    return new Promise((resolve) => {
+      resolve()
     })
   }
 
+  /**
+   * Inserts new records into a store.
+   * Must be implemented in a subclass.
+   *
+   * @param {object[]} records - An array of records to insert.
+   * @returns {Promise<undefined>|Promise<Error>} A promise that is resolved if records were inserted
+   *          successfully and is rejected if insertion failed.
+   */
   insert (records) {
-    return new Promise((resolve, reject) => {
-      if (!records) { resolve() } // Do nothing
-      if (!Array.isArray(records)) { records = [records] }
-      if (!this._db) reject(new Error('Database object is missing'))
-      const transaction = this._db.transaction(this._schema.name, Store.accessModes.READ_WRITE)
-      const store = transaction.objectStore(this._schema.name)
-      records.forEach(record => store.put(record))
-      transaction.oncomplete = () => resolve()
-      transaction.onerror = (error) => reject(error)
+    return new Promise((resolve) => {
+      resolve()
     })
   }
 
+  /**
+   * Updates records that already exist in a store.
+   * Must be implemented in a subclass.
+   *
+   * @param {[any, object]|[[any, object]]} keyValRecordsArr - A single item or an array of items
+   *        to insert. Each item is an array with key as a first member and a record to insert as a second one.
+   * @returns {Promise<undefined>|Promise<Error>} A promise that is resolved if records were updated
+   *          successfully and is rejected if operation failed.
+   */
   update (keyValRecordsArr) {
-    return new Promise((resolve, reject) => {
-      if (!keyValRecordsArr) resolve() // Do nothing
-      if (!Array.isArray(keyValRecordsArr)) reject(new Error('Record format must be [key,val] or [[key,val]]'))
-      if (!Array.isArray(keyValRecordsArr[0])) { keyValRecordsArr = [keyValRecordsArr] }
-      if (!this._db) reject(new Error('Database object is missing'))
-      const transaction = this._db.transaction(this._schema.name, Store.accessModes.READ_WRITE)
-      const store = transaction.objectStore(this._schema.name)
-      keyValRecordsArr.forEach(record => store.put(record[1], record[0]))
-      transaction.oncomplete = () => resolve()
-      transaction.onerror = (error) => reject(error)
+    return new Promise((resolve) => {
+      resolve()
     })
   }
 
+  /**
+   * Returns a total number of records in a store.
+   * Must be implemented in a subclass.
+   *
+   * @returns {Promise<number>|Promise<Error>} A promise that is resolved with number of records
+   *          in a store and is rejected if operation failed.
+   */
   count () {
     return new Promise((resolve) => {
-      if (!this._db) throw new Error('Database object is missing')
-      const transaction = this._db.transaction(this._schema.name, Store.accessModes.READ)
-      const store = transaction.objectStore(this._schema.name)
-      const countRequest = store.count()
-      countRequest.onsuccess = () => { resolve(countRequest.result) }
+      resolve()
     })
   }
 }
